modified:   qimgv/components/directorymanager/directorymanager.cpp

+++ qimgv/components/directorymanager/directorymanager.cpp
@@ -216,7 +216,8 @@ QString DirectoryManager::prevOfFile(QString filePath) const {
 QString DirectoryManager::nextOfFile(QString filePath) const {
     QString nextFilePath = "";
     int currentIndex = indexOfFile(filePath);
-    if(currentIndex >= 0 && currentIndex < fileEntryVec.size() - 1)
+    // 修复：将 size() 转换为 int 进行比较，避免 signed/unsigned 不匹配
+    if(currentIndex >= 0 && currentIndex < static_cast<int>(fileEntryVec.size()) - 1)
         nextFilePath = fileEntryVec.at(currentIndex + 1).path;
     return nextFilePath;
 }
@@ -232,7 +233,8 @@ QString DirectoryManager::prevOfDir(QString dirPath) const {
 QString DirectoryManager::nextOfDir(QString dirPath) const {
     QString nextDirectoryPath = "";
     int currentIndex = indexOfDir(dirPath);
-    if(currentIndex >= 0 && currentIndex < dirEntryVec.size() - 1)
+    // 修复：将 size() 转换为 int 进行比较
+    if(currentIndex >= 0 && currentIndex < static_cast<int>(dirEntryVec.size()) - 1)
         nextDirectoryPath = dirEntryVec.at(currentIndex + 1).path;
     return nextDirectoryPath;
 }
